Home page

+++ Backend/Movieverse.Domain/ValueObjects/BasicStatistics.cs
@@ -4,7 +4,7 @@ namespace Movieverse.Domain.ValueObjects;
 
 public sealed class BasicStatistics : ValueObject
 {
-	public short Rating { get; set; }
+	public double Rating { get; set; }
 	public int Votes { get; set; }
 	public int UserReviews { get; set; }
 	public int CriticReviews { get; set; }
